requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
supabase==2.3.4
stripe==7.8.0
razorpay==1.4.1
Werkzeug==3.0.1
gunicorn==21.2.0
requests==2.31.0
Flask-Session==0.6.0
